+++ zinnia/signals.py
@@ -86,17 +86,20 @@ def count_discussions_handler(sender, **kwargs):
         entry.comment_count = entry.comments.count()
         entry.pingback_count = entry.pingbacks.count()
         entry.trackback_count = entry.trackbacks.count()
-        entry.save(force_update=True)
+        entry.save(update_fields=[
+            'comment_count', 'pingback_count', 'trackback_count'])
 
 
 def count_comments_handler(sender, **kwargs):
     """
-    Update Entry.comment_count when a comment was posted.
+    Update Entry.comment_count when a public comment was posted.
     """
-    entry = kwargs['comment'].content_object
-    if isinstance(entry, Entry):
-        entry.comment_count = F('comment_count') + 1
-        entry.save(force_update=True)
+    comment = kwargs['comment']
+    if comment.is_public:
+        entry = comment.content_object
+        if isinstance(entry, Entry):
+            entry.comment_count = F('comment_count') + 1
+            entry.save(update_fields=['comment_count'])
 
 
 def count_pingbacks_handler(sender, **kwargs):
@@ -105,7 +108,7 @@ def count_pingbacks_handler(sender, **kwargs):
     """
     entry = kwargs['entry']
     entry.pingback_count = F('pingback_count') + 1
-    entry.save(force_update=True)
+    entry.save(update_fields=['pingback_count'])
 
 
 def count_trackbacks_handler(sender, **kwargs):
@@ -114,7 +117,7 @@ def count_trackbacks_handler(sender, **kwargs):
     """
     entry = kwargs['entry']
     entry.trackback_count = F('trackback_count') + 1
-    entry.save(force_update=True)
+    entry.save(update_fields=['trackback_count'])
 
 
 def connect_entry_signals():