+++ zinnia/tests/admin_filters.py
@@ -1,4 +1,6 @@
 """Test cases for Zinnia's admin filters"""
+from __future__ import unicode_literals
+
 from django.test import TestCase
 from django.test import RequestFactory
 from django.contrib.admin import site
@@ -7,6 +9,7 @@ from django.contrib.sites.models import Site
 from django.utils.translation import activate
 from django.utils.translation import deactivate
 from django.contrib.admin.views.main import ChangeList
+from django.contrib.auth.tests.utils import skipIfCustomUser
 
 from zinnia.models.entry import Entry
 from zinnia.models.author import Author
@@ -24,6 +27,7 @@ class MiniEntryCategoryAdmin(ModelAdmin):
     list_filter = [CategoryListFilter]
 
 
+@skipIfCustomUser
 class AuthorListFilterTestCase(TestCase):
     """Test case for AuthorListFilter"""
     urls = 'zinnia.tests.urls'
@@ -84,10 +88,10 @@ class AuthorListFilterTestCase(TestCase):
         with self.assertNumQueries(1):
             filterspec = changelist.get_filters(request)[0][0]
             self.assertEquals(filterspec.title, 'published authors')
-            self.assertEquals(filterspec.used_parameters, {'author': u'2'})
+            self.assertEquals(filterspec.used_parameters, {'author': '2'})
             self.assertEquals(filterspec.lookup_choices,
-                              [('1', u'webmaster (2 entries)'),
-                               ('2', u'contributor (1 entry)')])
+                              [('1', 'webmaster (2 entries)'),
+                               ('2', 'contributor (1 entry)')])
 
 
 class CategoryListFilterTestCase(TestCase):
@@ -150,7 +154,7 @@ class CategoryListFilterTestCase(TestCase):
         with self.assertNumQueries(1):
             filterspec = changelist.get_filters(request)[0][0]
             self.assertEquals(filterspec.title, 'published categories')
-            self.assertEquals(filterspec.used_parameters, {'category': u'2'})
+            self.assertEquals(filterspec.used_parameters, {'category': '2'})
             self.assertEquals(filterspec.lookup_choices,
-                              [('1', u'Category 1 (2 entries)'),
-                               ('2', u'Category 2 (1 entry)')])
+                              [('1', 'Category 1 (2 entries)'),
+                               ('2', 'Category 2 (1 entry)')])