+++ zinnia/tests/moderator.py
@@ -1,11 +1,15 @@
 """Test cases for Zinnia's moderator"""
+from __future__ import unicode_literals
+
 from django.core import mail
 from django.test import TestCase
+from django.utils import timezone
 from django.contrib import comments
 from django.contrib.sites.models import Site
 from django.core.urlresolvers import reverse
 from django.contrib.comments.forms import CommentForm
 from django.contrib.comments.moderation import moderator as moderator_stack
+from django.contrib.auth.tests.utils import skipIfCustomUser
 
 from zinnia.models.entry import Entry
 from zinnia.models.author import Author
@@ -15,6 +19,7 @@ from zinnia.signals import connect_discussion_signals
 from zinnia.signals import disconnect_discussion_signals
 
 
+@skipIfCustomUser
 class EntryCommentModeratorTestCase(TestCase):
     """Test cases for the moderator"""
 
@@ -33,7 +38,8 @@ class EntryCommentModeratorTestCase(TestCase):
     def test_email(self):
         comment = comments.get_model().objects.create(
             comment='My Comment', user=self.author, is_public=True,
-            content_object=self.entry, site=self.site)
+            content_object=self.entry, submit_date=timezone.now(),
+            site=self.site)
         self.assertEquals(len(mail.outbox), 0)
         moderator = EntryCommentModerator(Entry)
         moderator.email_reply = False
@@ -50,7 +56,8 @@ class EntryCommentModeratorTestCase(TestCase):
     def test_do_email_notification(self):
         comment = comments.get_model().objects.create(
             comment='My Comment', user=self.author, is_public=True,
-            content_object=self.entry, site=self.site)
+            content_object=self.entry, submit_date=timezone.now(),
+            site=self.site)
         self.assertEquals(len(mail.outbox), 0)
         moderator = EntryCommentModerator(Entry)
         moderator.mail_comment_notification_recipients = ['admin@example.com']
@@ -60,12 +67,13 @@ class EntryCommentModeratorTestCase(TestCase):
     def test_do_email_authors(self):
         comment = comments.get_model().objects.create(
             comment='My Comment', user=self.author, is_public=True,
-            content_object=self.entry, site=self.site)
+            content_object=self.entry, submit_date=timezone.now(),
+            site=self.site)
         self.assertEquals(len(mail.outbox), 0)
         moderator = EntryCommentModerator(Entry)
         moderator.email_authors = True
         moderator.mail_comment_notification_recipients = [
-            u'admin@example.com', u'webmaster@example.com']
+            'admin@example.com', 'webmaster@example.com']
         moderator.do_email_authors(comment, self.entry, 'request')
         self.assertEquals(len(mail.outbox), 0)
         moderator.mail_comment_notification_recipients = []
@@ -78,7 +86,8 @@ class EntryCommentModeratorTestCase(TestCase):
         """
         comment = comments.get_model().objects.create(
             comment='My Comment', user=self.author, is_public=True,
-            content_object=self.entry, site=self.site)
+            content_object=self.entry, submit_date=timezone.now(),
+            site=self.site)
         self.assertEquals(len(mail.outbox), 0)
         moderator = EntryCommentModerator(Entry)
         moderator.email_authors = True
@@ -89,50 +98,55 @@ class EntryCommentModeratorTestCase(TestCase):
         moderator.do_email_authors(comment, self.entry, 'request')
         self.assertEquals(len(mail.outbox), 1)
         self.assertEquals(mail.outbox[0].to,
-                          [u'admin@example.com', u'contrib@example.com'])
+                          ['admin@example.com', 'contrib@example.com'])
         mail.outbox = []
         contributor.email = ''
         contributor.save()
         moderator.do_email_authors(comment, self.entry, 'request')
         self.assertEquals(len(mail.outbox), 1)
-        self.assertEquals(mail.outbox[0].to, [u'admin@example.com'])
+        self.assertEquals(mail.outbox[0].to, ['admin@example.com'])
 
     def test_do_email_reply(self):
         comment = comments.get_model().objects.create(
             comment='My Comment 1', user=self.author, is_public=True,
-            content_object=self.entry, site=self.site)
+            content_object=self.entry, submit_date=timezone.now(),
+            site=self.site)
         moderator = EntryCommentModerator(Entry)
         moderator.email_notification_reply = True
         moderator.mail_comment_notification_recipients = [
-            u'admin@example.com', u'webmaster@example.com']
+            'admin@example.com', 'webmaster@example.com']
         moderator.do_email_reply(comment, self.entry, 'request')
         self.assertEquals(len(mail.outbox), 0)
 
         comment = comments.get_model().objects.create(
             comment='My Comment 2', user_email='user_1@example.com',
-            content_object=self.entry, is_public=True, site=self.site)
+            content_object=self.entry, is_public=True,
+            submit_date=timezone.now(), site=self.site)
         moderator.do_email_reply(comment, self.entry, 'request')
         self.assertEquals(len(mail.outbox), 0)
 
         comment = comments.get_model().objects.create(
             comment='My Comment 3', user_email='user_2@example.com',
-            content_object=self.entry, is_public=True, site=self.site)
+            content_object=self.entry, is_public=True,
+            submit_date=timezone.now(), site=self.site)
         moderator.do_email_reply(comment, self.entry, 'request')
         self.assertEquals(len(mail.outbox), 1)
-        self.assertEquals(mail.outbox[0].bcc, [u'user_1@example.com'])
+        self.assertEquals(mail.outbox[0].bcc, ['user_1@example.com'])
 
         comment = comments.get_model().objects.create(
             comment='My Comment 4', user=self.author, is_public=True,
-            content_object=self.entry, site=self.site)
+            content_object=self.entry, submit_date=timezone.now(),
+            site=self.site)
         moderator.do_email_reply(comment, self.entry, 'request')
         self.assertEquals(len(mail.outbox), 2)
-        self.assertEquals(mail.outbox[1].bcc, [u'user_1@example.com',
-                                               u'user_2@example.com'])
+        self.assertEquals(mail.outbox[1].bcc, ['user_1@example.com',
+                                               'user_2@example.com'])
 
     def test_moderate(self):
         comment = comments.get_model().objects.create(
             comment='My Comment', user=self.author, is_public=True,
-            content_object=self.entry, site=self.site)
+            content_object=self.entry, submit_date=timezone.now(),
+            site=self.site)
         moderator = EntryCommentModerator(Entry)
         moderator.auto_moderate_comments = True
         moderator.spam_checker_backends = ()
@@ -150,7 +164,8 @@ class EntryCommentModeratorTestCase(TestCase):
         self.entry.authors.clear()
         comment = comments.get_model().objects.create(
             comment='My Comment', user=self.author, is_public=True,
-            content_object=self.entry, site=self.site)
+            content_object=self.entry, submit_date=timezone.now(),
+            site=self.site)
         moderator = EntryCommentModerator(Entry)
         moderator.auto_moderate_comments = False
         moderator.spam_checker_backends = (
@@ -181,4 +196,4 @@ class EntryCommentModeratorTestCase(TestCase):
         disconnect_discussion_signals()
         self.assertEqual(comments.get_model().objects.count(), 1)
         entry = Entry.objects.get(pk=self.entry.pk)
-        self.assertEquals(entry.comment_count, 1)
+        self.assertEquals(entry.comment_count, 0)