+++ docs/notes/changelog.rst
@@ -4,6 +4,9 @@ CHANGELOG
 0.13
 ----
 
+* Python 3.0 support
+* Display page number in list
+* Basic support of custom User
 * Django 1.4 is no longer supported
 
 0.12.3